Move ClassWatcher to observer.js
